requirements.txt
psycopg2
python-dotenv
openai>=1.2.0